+++ src/main/java/entity/localEvent/LocalEventInterface.java
@@ -6,10 +6,23 @@ import java.time.LocalDate;
 
 import java.time.LocalTime;
 
+/**
+ * Interface for LocalEvent
+ * @see LocalEvent
+ */
 public interface LocalEventInterface extends NoteInterface {
+    /**
+     * @return the date of the event
+     */
     LocalDate getDate();
 
+    /**
+     * @return the start time of the event
+     */
     LocalTime getStartTime();
 
+    /**
+     * @return the end time of the event
+     */
     LocalTime getEndTime();
 }
