+++ src/main/java/interface_adapter/localEvent/LocalEventPresenter.java
@@ -11,6 +11,9 @@ import java.util.ArrayList;
 
 import java.util.HashMap;
 
+/**
+ * Presenter for local events.
+ */
 public class LocalEventPresenter implements LocalEventOutputBoundary {
     private final LocalEventViewModel detailedLocalEventViewModel;
 
@@ -22,6 +25,15 @@ public class LocalEventPresenter implements LocalEventOutputBoundary {
 
     private final ViewManagerModel viewManagerModel;
 
+    /**
+     * Constructor for LocalEventPresenter.
+     *
+     * @param viewManagerModel             The view manager model.
+     * @param detailedLocalEventViewModel  The detailed local event view model.
+     * @param localEventCreationViewModel  The local event creation view model.
+     * @param localEventEditViewModel      The local event edit view model.
+     * @param listViewModel                The list view model.
+     */
     public LocalEventPresenter(ViewManagerModel viewManagerModel,
                                LocalEventViewModel detailedLocalEventViewModel,
                                LocalEventViewModel localEventCreationViewModel,
@@ -34,6 +46,12 @@ public class LocalEventPresenter implements LocalEventOutputBoundary {
         this.listViewModel = listViewModel;
     }
 
+    /**
+     * Creates a local event state.
+     *
+     * @param localEventOutputData  The local event output data.
+     * @return                      The local event state.
+     */
     public LocalEventState CreateState(LocalEventOutputData localEventOutputData) {
         return new LocalEventState(localEventOutputData.getTitle(),
                 localEventOutputData.getLocation(),
@@ -46,6 +64,11 @@ public class LocalEventPresenter implements LocalEventOutputBoundary {
                 localEventOutputData.getEndTime());
     }
 
+    /**
+     * Updates the list of events in the list view model and displays the
+     * event's detailed view.
+     * @param localEventOutputData The local event output data.
+     */
     @Override
     public void UpdateEventsList(LocalEventOutputData
                                                    localEventOutputData) {
@@ -59,6 +82,11 @@ public class LocalEventPresenter implements LocalEventOutputBoundary {
         DisplayEventDetailedView(localEventOutputData);
     }
 
+    /**
+     * Updates the list of events in the list view model after deleting an
+     * event and switches to the list view.
+     * @param ID The ID of the event to be deleted.
+     */
     @Override
     public void DeleteEventSuccessView(int ID) {
         ListViewState listViewState = listViewModel.getState();
@@ -70,11 +98,19 @@ public class LocalEventPresenter implements LocalEventOutputBoundary {
         viewManagerModel.firePropertyChanged();
     }
 
+    /**
+     * Displays an error message.
+     * @param errorMessage The error message.
+     */
     @Override
     public void EventFailView(String errorMessage) {
         System.out.println(errorMessage);
     }
 
+    /**
+     * Displays the detailed view of an event.
+     * @param localEventOutputData The local event output data.
+     */
     @Override
     public void DisplayEventDetailedView(LocalEventOutputData localEventOutputData) {
         LocalEventState localEventState = CreateState(localEventOutputData);
@@ -85,6 +121,9 @@ public class LocalEventPresenter implements LocalEventOutputBoundary {
         viewManagerModel.firePropertyChanged();
     }
 
+    /**
+     * Displays the view for creating a new event.
+     */
     @Override
     public void DisplayEventCreationView() {
         LocalEventState localEventState = new LocalEventState();
@@ -95,6 +134,10 @@ public class LocalEventPresenter implements LocalEventOutputBoundary {
         viewManagerModel.firePropertyChanged();
     }
 
+    /**
+     * Displays the view for editing an event.
+     * @param eventOutputData The local event output data.
+     */
     @Override
     public void DisplayEventEditView(LocalEventOutputData eventOutputData) {
         LocalEventState localEventState = CreateState(eventOutputData);
@@ -104,11 +147,14 @@ public class LocalEventPresenter implements LocalEventOutputBoundary {
         viewManagerModel.firePropertyChanged();
     }
 
+    /**
+     * Displays the list view of events.
+     * @param localEventOutputDataArrayList The list of local event output data.
+     */
     @Override
     public void DisplayAllEvents(ArrayList<LocalEventOutputData>
                                              localEventOutputDataArrayList) {
-        HashMap<Integer, HashMap<String, ?>> listViewEvents =
-                new HashMap<Integer, HashMap<String, ?>>();
+        HashMap<Integer, HashMap<String, ?>> listViewEvents = new HashMap<>();
         for (LocalEventOutputData localEventOutputData :
                 localEventOutputDataArrayList) {
             listViewEvents.put(localEventOutputData.getID(),
