+++ src/main/java/interface_adapter/localEvent/LocalEventState.java
@@ -8,6 +8,9 @@ import java.time.LocalTime;
 
 import java.util.ArrayList;
 
+/**
+ * State for local events.
+ */
 public class LocalEventState extends NoteState {
     private LocalDate date;
 
@@ -18,6 +21,19 @@ public class LocalEventState extends NoteState {
     public LocalEventState() {
     }
 
+    /**
+     * Constructor for LocalEventState.
+     *
+     * @param title         The title.
+     * @param location      The location.
+     * @param description   The description.
+     * @param isWork        Whether the local event is work.
+     * @param pinned        Whether the local event is pinned.
+     * @param subEvents     The sub events.
+     * @param date          The date.
+     * @param startTime     The start time.
+     * @param endTime       The end time.
+     */
     public LocalEventState(String title, String location, String description,
                            boolean isWork, boolean pinned,
                            ArrayList<Integer> subEvents, LocalDate date,
@@ -31,14 +47,29 @@ public class LocalEventState extends NoteState {
         this.endTime = endTime;
     }
 
+    /**
+     * Gets the date.
+     *
+     * @return  The date.
+     */
     public LocalDate getDate() {
         return date;
     }
 
+    /**
+     * Gets the start time.
+     *
+     * @return  The start time.
+     */
     public LocalTime getStartTime() {
         return startTime;
     }
 
+    /**
+     * Gets the end time.
+     *
+     * @return  The end time.
+     */
     public LocalTime getEndTime() {
         return endTime;
     }
