+++ src/main/java/interface_adapter/localEvent/LocalEventViewModel.java
@@ -6,6 +6,11 @@ import java.beans.PropertyChangeListener;
 
 import java.beans.PropertyChangeSupport;
 
+/**
+ * This class is a ViewModel for the LocalEventView.
+ * It contains the state of the view.
+ * It also contains the labels for the view.
+ */
 public class LocalEventViewModel extends ViewModel {
     public static final String TITLE_VIEW = "Event view";
 
@@ -29,14 +34,25 @@ public class LocalEventViewModel extends ViewModel {
 
     private LocalEventState state = new LocalEventState();
 
+    /**
+     * Constructor for the LocalEventViewModel.
+     * @param viewName The name of the view.
+     */
     public LocalEventViewModel(String viewName) {
         super(viewName);
     }
 
+    /**
+     * Constructor for the LocalEventViewModel.
+     */
     public void setState(LocalEventState state) {
         this.state = state;
     }
 
+    /**
+     * Getter for the state of the view.
+     * @return The state of the view.
+     */
     public LocalEventState getState() {
         return state;
     }
@@ -44,11 +60,18 @@ public class LocalEventViewModel extends ViewModel {
     private final PropertyChangeSupport support = new PropertyChangeSupport(
             this);
 
+    /**
+     * Fires a property change event.
+     */
     public void firePropertyChanged() {
         support.firePropertyChange("state", null,
                 this.state);
     }
 
+    /**
+     * Adds a property change listener.
+     * @param listener The listener to be added.
+     */
     public void addPropertyChangeListener(PropertyChangeListener listener) {
         support.addPropertyChangeListener(listener);
     }
