+++ src/main/java/interface_adapter/note/NotePresenter.java
@@ -11,6 +11,9 @@ import java.util.ArrayList;
 
 import java.util.HashMap;
 
+/**
+ * This class is a Presenter for the NoteView.
+ */
 public class NotePresenter implements NoteOutputBoundary {
     private final NoteViewModel detailedNoteViewModel;
 
@@ -22,6 +25,15 @@ public class NotePresenter implements NoteOutputBoundary {
 
     private final ViewManagerModel viewManagerModel;
 
+    /**
+     * Constructor for NotePresenter.
+     *
+     * @param viewManagerModel         The view manager model.
+     * @param detailedNoteViewModel    The detailed note view model.
+     * @param noteCreationViewModel    The note creation view model.
+     * @param noteEditViewModel        The note edit view model.
+     * @param listViewModel            The list view model.
+     */
     public NotePresenter(ViewManagerModel viewManagerModel,
                          NoteViewModel detailedNoteViewModel,
                          NoteViewModel noteCreationViewModel,
@@ -34,6 +46,12 @@ public class NotePresenter implements NoteOutputBoundary {
         this.listViewModel = listViewModel;
     }
 
+    /**
+     * Creates a note state.
+     *
+     * @param noteOutputData  The note output data.
+     * @return                The note state.
+     */
     public NoteState CreateState(NoteOutputData noteOutputData) {
         return new NoteState(noteOutputData.getTitle(),
                 noteOutputData.getLocation(),
@@ -43,6 +61,11 @@ public class NotePresenter implements NoteOutputBoundary {
                 noteOutputData.getSubEvents());
     }
 
+    /**
+     * Updates the list view's notes list with the given note output data and
+     * switches to a detailed view of the note.
+     * @param noteOutputData The note output data.
+     */
     @Override
     public void UpdateNotesList(NoteOutputData
                                          noteOutputData) {
@@ -56,6 +79,11 @@ public class NotePresenter implements NoteOutputBoundary {
         DisplayNoteDetailedView(noteOutputData);
     }
 
+    /**
+     * Delete the note with the given ID from the list view's notes list and
+     * switch to the list view.
+     * @param ID The ID of the note to be deleted.
+     */
     @Override
     public void DeleteNoteSuccessView(int ID) {
         ListViewState listViewState = listViewModel.getState();
@@ -67,11 +95,19 @@ public class NotePresenter implements NoteOutputBoundary {
         viewManagerModel.firePropertyChanged();
     }
 
+    /**
+     * Displays an error message.
+     * @param errorMessage The error message.
+     */
     @Override
     public void NoteFailView(String errorMessage) {
         System.out.println(errorMessage);
     }
 
+    /**
+     * Displays a detailed view of the note with the given note output data.
+     * @param noteOutputData The note output data.
+     */
     @Override
     public void DisplayNoteDetailedView(NoteOutputData noteOutputData) {
         NoteState noteState = new NoteState(
@@ -88,6 +124,9 @@ public class NotePresenter implements NoteOutputBoundary {
         viewManagerModel.firePropertyChanged();
     }
 
+    /**
+     * Displays the note creation view.
+     */
     @Override
     public void DisplayNoteCreationView() {
         NoteState noteState = new NoteState();
@@ -98,6 +137,10 @@ public class NotePresenter implements NoteOutputBoundary {
         viewManagerModel.firePropertyChanged();
     }
 
+    /**
+     * Displays the note edit view with the given note output data.
+     * @param noteOutputData The note output data.
+     */
     @Override
     public void DisplayNoteEditView(NoteOutputData noteOutputData) {
         NoteState noteState = CreateState(noteOutputData);
@@ -107,6 +150,10 @@ public class NotePresenter implements NoteOutputBoundary {
         viewManagerModel.firePropertyChanged();
     }
 
+    /**
+     * Displays all notes in the array in list view.
+     * @param noteOutputDataArrayList The note output data array list.
+     */
     @Override
     public void DisplayAllNotes(ArrayList<NoteOutputData>
                                          noteOutputDataArrayList) {
