+++ src/main/java/interface_adapter/note/NoteViewModel.java
@@ -6,6 +6,11 @@ import java.beans.PropertyChangeListener;
 
 import java.beans.PropertyChangeSupport;
 
+/**
+ * This class is a ViewModel for the NoteView.
+ * It contains the state of the view.
+ * It also contains the labels for the view.
+ */
 public class NoteViewModel extends ViewModel {
     public static final String TITLE_VIEW = "Note view";
 
@@ -23,14 +28,25 @@ public class NoteViewModel extends ViewModel {
 
     private NoteState state = new NoteState();
 
+    /**
+     * Constructor for the NoteViewModel.
+     * @param viewName The name of the view.
+     */
     public NoteViewModel(String viewName) {
         super(viewName);
     }
 
+    /**
+     * Setter for the state of the view.
+     */
     public void setState(NoteState state) {
         this.state = state;
     }
 
+    /**
+     * Getter for the state of the view.
+     * @return The state of the view.
+     */
     public NoteState getState() {
         return state;
     }
@@ -38,11 +54,18 @@ public class NoteViewModel extends ViewModel {
     private final PropertyChangeSupport support = new PropertyChangeSupport(
             this);
 
+    /**
+     * Fires a property change.
+     */
     public void firePropertyChanged() {
         support.firePropertyChange("state", null,
                 this.state);
     }
 
+    /**
+     * Adds a property change listener.
+     * @param listener The property change listener.
+     */
     public void addPropertyChangeListener(PropertyChangeListener listener) {
         support.addPropertyChangeListener(listener);
     }
