+++ src/main/java/use_case/localEvent/LocalEventOutputData.java
@@ -9,6 +9,9 @@ import java.time.LocalTime;
 import java.util.ArrayList;
 import java.util.HashMap;
 
+/**
+ * The output data for local events
+ */
 public class LocalEventOutputData extends NoteOutputData {
     private final LocalDate date;
     // default startTime when user specifies whole-day event
@@ -17,6 +20,20 @@ public class LocalEventOutputData extends NoteOutputData {
 
     private LocalTime endTime = LocalTime.of(23, 59);
 
+    /**
+     * Creates a new LocalEventOutputData object
+     * @param ID The ID of the event
+     * @param title The title of the event
+     * @param userID The ID of the user who created the event
+     * @param date The date of the event
+     * @param startTime The start time of the event
+     * @param endTime The end time of the event
+     * @param location The location of the event
+     * @param description The description of the event
+     * @param isWork Whether the event is a work event
+     * @param pinned Whether the event is pinned
+     * @param subEvents The IDs of the sub-events of the event
+     */
     public LocalEventOutputData(int ID, String title, int userID, LocalDate date,
                                 LocalTime startTime, LocalTime endTime,
                                 String location, String description, boolean isWork,
@@ -31,6 +48,10 @@ public class LocalEventOutputData extends NoteOutputData {
         this.endTime = endTime;
     }
 
+    /**
+     * Returns a dictionary of the event data
+     * @return The dictionary of the event data
+     */
     public HashMap<String, Object> getEventData() {
         HashMap<String, Object> data = super.getNoteData();
         data.put("date", getDate());
@@ -39,14 +60,26 @@ public class LocalEventOutputData extends NoteOutputData {
         return data;
     }
 
+    /**
+     * Returns the date of the event
+     * @return The date of the event
+     */
     public LocalDate getDate() {
         return date;
     }
 
+    /**
+     * Returns the start time of the event
+     * @return The start time of the event
+     */
     public LocalTime getStartTime() {
         return startTime;
     }
 
+    /**
+     * Returns the end time of the event
+     * @return The end time of the event
+     */
     public LocalTime getEndTime() {
         return endTime;
     }
