+++ src/main/java/use_case/note/NoteInteractor.java
@@ -1,7 +1,5 @@
 package use_case.note;
 
-import InputData.NoteInputData;
-
 import data_access.NotesDataAccessObject;
 
 import entity.NoteFactoryInterface;
@@ -10,6 +8,9 @@ import entity.note.NoteFactory;
 
 import java.util.ArrayList;
 
+/**
+ * The interactor for the Note entity
+ */
 public class NoteInteractor implements NoteInputBoundary {
     private final NoteOutputBoundary noteOutputBoundary;
 
@@ -17,6 +18,11 @@ public class NoteInteractor implements NoteInputBoundary {
 
     private final NoteFactoryInterface noteFactory;
 
+    /**
+     * Constructor for NoteInteractor
+     * @param noteOutputBoundary The presenter for the NoteInteractor
+     * @param notesDataAccessObject The data access object for the NoteInteractor
+     */
     public NoteInteractor(NoteOutputBoundary noteOutputBoundary,
                           NotesDataAccessObject<Note> notesDataAccessObject){
         this.noteOutputBoundary = noteOutputBoundary;
@@ -26,10 +32,18 @@ public class NoteInteractor implements NoteInputBoundary {
         this.noteFactory = new NoteFactory();
     }
 
+    /**
+     * Displays the note creation view
+     */
+    @Override
     public void DisplayNoteCreationView() {
         noteOutputBoundary.DisplayNoteCreationView();
     }
 
+    /**
+     * Displays the note edit view
+     * @param noteID The ID of the note to be edited
+     */
     @Override
     public void DisplayNoteEditView(int noteID) {
         Note note = notesDataAccessObject.getByID(noteID);
@@ -40,6 +54,10 @@ public class NoteInteractor implements NoteInputBoundary {
                 note.getPinned(), note.getDescendants()));
     }
 
+    /**
+     * Displays the note detailed view
+     * @param noteID The ID of the note to be displayed
+     */
     @Override
     public void DisplayNoteDetailedView(int noteID) {
         Note note = notesDataAccessObject.getByID(noteID);
@@ -50,6 +68,10 @@ public class NoteInteractor implements NoteInputBoundary {
                 note.getPinned(), note.getDescendants()));
     }
 
+    /**
+     * Creates a note and saves it to the database
+     * @param noteInputData The input data for the note to be created
+     */
     @Override
     public void CreateNote(NoteInputData noteInputData) {
         try {
@@ -73,6 +95,10 @@ public class NoteInteractor implements NoteInputBoundary {
         }
     }
 
+    /**
+     * Edits a note and saves it to the database
+     * @param noteInputData The input data for the note to be edited
+     */
     @Override
     public void EditNote(NoteInputData noteInputData) {
         try {
@@ -98,6 +124,10 @@ public class NoteInteractor implements NoteInputBoundary {
         }
     }
 
+    /**
+     * Deletes a note from the database
+     * @param noteID The ID of the note to be deleted
+     */
     @Override
     public void DeleteNote(int noteID) {
         try {
@@ -112,6 +142,10 @@ public class NoteInteractor implements NoteInputBoundary {
         }
     }
 
+    /**
+     * Displays all notes in the database
+     * @param userID The ID of the user whose notes are to be displayed
+     */
     @Override
     public void GetAllNotes(int userID) {
         ArrayList<NoteOutputData> noteOutputDataArrayList = new ArrayList<>();
